+++ CHANGELOG.md
@@ -3,6 +3,20 @@
 This file summarises the most important changes that went live on each release
 of Concierge. Please check the Wiki entry on the release process to understand
 how this file is formatted and how the process works.
+
+## Unreleased
+-
+
+## [0.13.2]
+### Fixed
+- Kigo: prevent context hell for availabilities worker
+- net_rate calculation
+- Rentals United check in time takes CheckInTo
+
+### Changed
+- Improved Ciirus context info during sync process
+- Avantio filter users by Login GA instead of GA Code
+
 ## [0.13.1] - 2016-11-08
 ## Changed
 - Cancellation policies naming